Mise en place de la sécurité avec la création de token en JWT

+++ controller/articleController.js
@@ -5,14 +5,20 @@ var router = require("express").Router();
 var bodyParser = require('body-parser');
 var multer = require('multer');
 var mime = require('mime');
+var morgan = require('morgan');
+var session = require('express-session');
+
 var Article = require('./../model/article');
 var articleMapper = require("./../helper/mapper/articleMapper");
 var articleService = require("./../service/articleService");
-var session = require('express-session');
+var env = require('./../config/environnementconf');
+var security = require('./../config/security');
+var jwt = require('./../helper/tokenHelper');
 
 var log = require('./../config/log4js').getLogger('gonnetLogger');
 
 router.use(bodyParser.json());
+router.use(morgan(env.environnement));
 
 log.level = 'debug';
 
@@ -33,51 +39,30 @@ var upload = multer({ //multer settings
     storage: storage
 });
 
-router.get('/' , function (req , res) {
+router.get('/:token' , function (req , res) {
     log.info('Article get');
     try{
-        var promise = articleService.getAllArticles();
-        promise.then(function(articles){
-            res.send(articles);
-            res.end();
-        });
-    }
-    catch(error)
-    {
-        log.error(error);
-    }
-});
-
-router.get('/:idArticle' , function (req , res) {
-    log.info('Article get with params');
-    try{
-        var promise = articleService.getArticle(req.params.idArticle);
-        promise.then(function (article) {
-            console.log(article);
-            res.send(article);
-            res.end();
-        }).catch(function (error) {
-            console.log(error);
-            log.error(error);
-        });
-    }
-    catch(error)
-    {
-        log.error(error);
-    }
-});
-
-router.put('/' ,upload.single('file'), function (req , res) {
-    log.info('Article POST');
-    try {
-        if(req.file != undefined)
+        var token = req.params.token;
+        if(token)
         {
-            console.log(req.file.filename);
-            console.log("2");
-            var article = articleMapper.createArticle(req, filename);
-            var result = articleService.createArticle(article);
-            res.send(result);
-            res.end();
+            var tokenVerification = jwt.verifyToken(token);
+            if(tokenVerification)
+            {
+                var promise = articleService.getAllArticles();
+                promise.then(function(articles){
+                    res.json({
+                        success : true,
+                        articles : articles
+                    });
+                    res.end();
+                });
+            }
+            else{
+                res.json({
+                    success : false
+                });
+                res.end();
+            }
         }
     }
     catch(error)
@@ -86,14 +71,78 @@ router.put('/' ,upload.single('file'), function (req , res) {
     }
 });
 
-router.delete('/' , function (req , res) {
-    articleService.deleteArticle(req, res);
+router.get('/:idArticle/:token' , function (req , res) {
+    log.info('Article get with params');
+    try{
+        var token = req.params.token;
+        if(token)
+        {
+            var tokenVerification = jwt.verifyToken(token);
+            if(tokenVerification)
+            {
+                var promise = articleService.getArticle(req.params.idArticle);
+                promise.then(function (article) {
+                    console.log(article);
+                    res.json({
+                        success : true,
+                        article : article
+                    });
+                    res.end();
+                }).catch(function (error) {
+                    console.log(error);
+                    log.error(error);
+                    res.json({
+                        success : false
+                    });
+                    res.end();
+                });
+            }
+        }
+
+    }
+    catch(error)
+    {
+        log.error(error);
+    }
 });
 
-router.post('/:idArticle' , upload.single('file') , function (req , res) {
+router.put('/:token' ,upload.single('file'), function (req , res) {
+    log.info('Article POST');
+    try {
+        if(req.file != undefined)
+        {
+            var tokenVerification = jwt.verifyToken(req.params.token);
+            if(tokenVerification)
+            {
+                var article = articleMapper.createArticle(req, filename);
+                var result = articleService.createArticle(article);
+                res.send(result);
+                res.end();
+            }
+        }
+    }
+    catch(error)
+    {
+        log.error(error);
+    }
+});
+
+router.delete('/:token' , function (req , res) {
+    var tokenVerification = jwt.verifyToken(req.params.token);
+    if(tokenVerification)
+    {
+        articleService.deleteArticle(req, res);
+    }
+});
+
+router.post('/:idArticle/:token' , upload.single('file') , function (req , res) {
     log.info('Article PUT');
     try{
-        var article = articleService.updateArticle(req , res);
+        var tokenVerification = jwt.verifyToken(req.params.token);
+        if(tokenVerification)
+        {
+            articleService.updateArticle(req , res);
+        }
     }
     catch(error)
     {
